Change the API to export the constructor directly.
This is more idiomatic for node modules, and avoids the
need for consumer to use two names in referring to the module
and constructor function seperately.

For now, also continues to export the constructore as a
member named `NodeAppInsights` off of the module (which is
the same constructor function itself) for back compat.

Tests now use the new API, but validate the existance of the
legacy API.