This closes #14, and closes #16 by refactoring the ip parser
In node v12 the behavior of request.connection.remoteAddress changed such that it returns dual stack by default. To mitigate this the IP is extracted via regex. One other side effect of this is that "::1" is returned when running in localhost. This could default to empty string and the payload would be valud, but setting it to "127.0.0.1" since that is probably more helpful.

This fixes #16 by stripping off the ipv6 part of remoteAddress which looks like "::ffff:127.0.0.1" in node v12. It fixes #14 because the port is also removed from the output by this regex.